Adds lazy-power-select
This commit adds the bootstrap lazy power select to allow the select to
lazy load paginated content. It also refactors the markup into a higher
order component called `bootstrap/control-wrapper` so it is easier to
reuse the bootstrap wrapping code and label. Eventually all input
components other than checkbox should probably be rewritten to use this.
It isn't completely backwards compatible though and requires a `label`
and `srOnly` flag to hide the label. Some current bootstrap inputs have
optional labels and I did not want to break this on people. We need to
deprecate label and eventually move all components to use it. This also
configures the annotation for this addon inside freestyle.